Move nativeapp feature flags to Snowflake CLI feature flags
