--- src/snowflake/cli/_plugins/nativeapp/feature_flags.py
@@ -1,28 +1,0 @@
-# Copyright (c) 2024 Snowflake Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from enum import unique
-
-from snowflake.cli.api.feature_flags import BooleanFlag, FeatureFlagMixin
-
-
-@unique
-class FeatureFlag(
-    FeatureFlagMixin
-):  # TODO move this to snowflake.cli.api.feature_flags
-    ENABLE_NATIVE_APP_PYTHON_SETUP = BooleanFlag(
-        "ENABLE_NATIVE_APP_PYTHON_SETUP", False
-    )
-    ENABLE_NATIVE_APP_CHILDREN = BooleanFlag("ENABLE_NATIVE_APP_CHILDREN", False)
-    ENABLE_RELEASE_CHANNELS = BooleanFlag("ENABLE_RELEASE_CHANNELS", None)
