+++ src/snowflake/cli/api/feature_flags.py
@@ -69,3 +69,8 @@ class FeatureFlag(FeatureFlagMixin):
     ENABLE_SNOWPARK_GLOB_SUPPORT = BooleanFlag("ENABLE_SNOWPARK_GLOB_SUPPORT", False)
     ENABLE_SPCS_SERVICE_EVENTS = BooleanFlag("ENABLE_SPCS_SERVICE_EVENTS", False)
     ENABLE_AUTH_KEYPAIR = BooleanFlag("ENABLE_AUTH_KEYPAIR", False)
+    ENABLE_NATIVE_APP_PYTHON_SETUP = BooleanFlag(
+        "ENABLE_NATIVE_APP_PYTHON_SETUP", False
+    )
+    ENABLE_NATIVE_APP_CHILDREN = BooleanFlag("ENABLE_NATIVE_APP_CHILDREN", False)
+    ENABLE_RELEASE_CHANNELS = BooleanFlag("ENABLE_RELEASE_CHANNELS", None)
