+++ tests_integration/nativeapp/test_publish.py
@@ -19,7 +19,7 @@ from tests_integration.nativeapp.native_apps_utils import get_org_and_account_na
 
 
 @mock.patch(
-    "snowflake.cli._plugins.nativeapp.feature_flags.FeatureFlag.ENABLE_RELEASE_CHANNELS.get_value",
+    "snowflake.cli.api.feature_flags.FeatureFlag.ENABLE_RELEASE_CHANNELS.get_value",
 )
 @pytest.mark.parametrize("release_channels_enabled", [True, False])
 @pytest.mark.integration
@@ -83,7 +83,7 @@ def test_publish_with_default_release_directive(
 
 
 @mock.patch(
-    "snowflake.cli._plugins.nativeapp.feature_flags.FeatureFlag.ENABLE_RELEASE_CHANNELS.get_value",
+    "snowflake.cli.api.feature_flags.FeatureFlag.ENABLE_RELEASE_CHANNELS.get_value",
 )
 @pytest.mark.parametrize("release_channels_enabled", [True, False])
 @pytest.mark.integration
@@ -171,7 +171,7 @@ def test_publish_with_existing_version_and_patch(
 
 
 @mock.patch(
-    "snowflake.cli._plugins.nativeapp.feature_flags.FeatureFlag.ENABLE_RELEASE_CHANNELS.get_value",
+    "snowflake.cli.api.feature_flags.FeatureFlag.ENABLE_RELEASE_CHANNELS.get_value",
     return_value=True,
 )
 @pytest.mark.integration
@@ -261,7 +261,7 @@ def test_publish_with_non_default_release_channel(
 
 
 @mock.patch(
-    "snowflake.cli._plugins.nativeapp.feature_flags.FeatureFlag.ENABLE_RELEASE_CHANNELS.get_value",
+    "snowflake.cli.api.feature_flags.FeatureFlag.ENABLE_RELEASE_CHANNELS.get_value",
 )
 @pytest.mark.parametrize("release_channels_enabled", [True, False])
 @pytest.mark.integration
